+++ R/adjust.R
@@ -22,12 +22,19 @@
 #'   effects.
 #' @param bayesian If `TRUE`, the models are fitted under the Bayesian framework
 #'   using `rstanarm`.
+#' @param keep_intercept If `FALSE` (default), the intercept of the model is
+#'   re-added. This avoids the centering around 0 that happens by default
+#'   when regressing out another variable (see the examples below for a
+#'   visual representation of this).
 #'
 #' @return A data frame comparable to `data`, with adjusted variables.
 #'
 #' @examples
-#' adjust(attitude)
-#' adjust(attitude, effect = "complaints", select = "rating")
+#' adjusted_all <- adjust(attitude)
+#' head(adjusted_all)
+#' adjusted_one <- adjust(attitude, effect = "complaints", select = "rating")
+#' head(adjusted_one)
+#'
 #' \donttest{
 #' \dontrun{
 #' adjust(attitude, effect = "complaints", select = "rating", bayesian = TRUE)
@@ -36,9 +43,28 @@
 #' adjust(attitude, effect = "complaints_LMH", select = "rating", multilevel = TRUE)
 #' }
 #' }
+#' if(require("bayestestR")){
+#' # Generate data
+#' data <- bayestestR::simulate_correlation(n=100, r=0.7)
+#' data$V2 <- (5 * data$V2) + 20  # Add intercept
 #'
+#' # Adjust
+#' adjusted <- adjust(data, effect="V1", select="V2")
+#' adjusted_icpt <- adjust(data, effect="V1", select="V2", keep_intercept=TRUE)
+#'
+#' # Visualize
+#' plot(data$V1, data$V2, pch = 19, col = "blue",
+#'   ylim=c(min(adjusted$V2), max(data$V2)),
+#'   main = "Original (blue), adjusted (green), and adjusted - intercept kept (red) data")
+#' abline(lm(V2 ~ V1, data = data), col = "blue")
+#' points(adjusted$V1, adjusted$V2, pch = 19, col = "green")
+#' abline(lm(V2 ~ V1, data = adjusted), col = "green")
+#' points(adjusted_icpt$V1, adjusted_icpt$V2, pch = 19, col = "red")
+#' abline(lm(V2 ~ V1, data = adjusted_icpt), col = "red")
+#' }
+#' @importFrom stats as.formula
 #' @export
-adjust <- function(data, effect = NULL, select = NULL, exclude = NULL, multilevel = FALSE, additive = FALSE, bayesian = FALSE) {
+adjust <- function(data, effect = NULL, select = NULL, exclude = NULL, multilevel = FALSE, additive = FALSE, bayesian = FALSE, keep_intercept = FALSE) {
   if (!all(colnames(data) == make.names(colnames(data), unique = TRUE))) {
     warning("Bad column names (e.g., with spaces) have been detected which might create issues in many functions.\n",
       "Please fix it (you can run `names(mydata) <- make.names(names(mydata))` for a quick fix).",
@@ -68,7 +94,7 @@ adjust <- function(data, effect = NULL, select = NULL, exclude = NULL, multileve
   if (length(facs) >= 1) {
     if (multilevel) {
       if (additive) {
-        formula_random <- as.formula(paste("~", paste(paste0("(1|", facs, ")"), collapse = " + ")))
+        formula_random <- stats::as.formula(paste("~", paste(paste0("(1|", facs, ")"), collapse = " + ")))
       } else {
         formula_random <- paste("+", paste(paste0("(1|", facs, ")"), collapse = " + "))
       }
@@ -96,19 +122,22 @@ adjust <- function(data, effect = NULL, select = NULL, exclude = NULL, multileve
     formula_predictors <- paste(c("1", predictors), collapse = " + ")
     formula <- paste(var, "~", formula_predictors)
 
-    x <- .model_adjust_for(data = data[unique(c(var, effect, facs))], formula, multilevel = multilevel, additive = additive, bayesian = bayesian, formula_random = formula_random)
+    x <- .model_adjust_for(data = data[unique(c(var, effect, facs))], formula, multilevel = multilevel, additive = additive, bayesian = bayesian, formula_random = formula_random, keep_intercept = keep_intercept)
     out[var] <- x
   }
   out[names(data)[!names(data) %in% names(out)]] <- data[names(data)[!names(data) %in% names(out)]]
   out[names(data)]
 }
 
+#' @rdname adjust
+#' @export
+data_adjust <- adjust
 
 
-
-#' @importFrom stats lm residuals as.formula complete.cases
+#' @importFrom stats lm residuals as.formula complete.cases median
+#' @importFrom insight get_intercept get_residuals
 #' @keywords internal
-.model_adjust_for <- function(data, formula, multilevel = FALSE, additive = FALSE, bayesian = FALSE, formula_random = NULL) {
+.model_adjust_for <- function(data, formula, multilevel = FALSE, additive = FALSE, bayesian = FALSE, formula_random = NULL, keep_intercept = FALSE) {
 
   # Additive -----------------------
   if (additive) {
@@ -117,13 +146,13 @@ adjust <- function(data, effect = NULL, select = NULL, exclude = NULL, multileve
       if (!requireNamespace("rstanarm")) {
         stop("This function needs `rstanarm` to be installed. Please install by running `install.packages('rstanarm')`.")
       }
-      adjusted <- rstanarm::stan_gamm4(as.formula(formula), random = formula_random, data = data, refresh = 0)$residuals
+      model <- rstanarm::stan_gamm4(stats::as.formula(formula), random = formula_random, data = data, refresh = 0)
       # Frequentist
     } else {
       if (!requireNamespace("gamm4")) {
         stop("This function needs `gamm4` to be installed. Please install by running `install.packages('gamm4')`.")
       }
-      adjusted <- gamm4::gamm4(as.formula(formula), random = formula_random, data = data)$gam$residuals
+      model <- gamm4::gamm4(stats::as.formula(formula), random = formula_random, data = data)
     }
 
     # Linear -------------------------
@@ -134,9 +163,9 @@ adjust <- function(data, effect = NULL, select = NULL, exclude = NULL, multileve
         stop("This function needs `rstanarm` to be installed. Please install by running `install.packages('rstanarm')`.")
       }
       if (multilevel) {
-        adjusted <- rstanarm::stan_lmer(paste(formula, formula_random), data = data, refresh = 0)$residuals
+        model <- rstanarm::stan_lmer(paste(formula, formula_random), data = data, refresh = 0)
       } else {
-        adjusted <- rstanarm::stan_glm(formula, data = data, refresh = 0)$residuals
+        model <- rstanarm::stan_glm(formula, data = data, refresh = 0)
       }
       # Frequentist
     } else {
@@ -144,11 +173,21 @@ adjust <- function(data, effect = NULL, select = NULL, exclude = NULL, multileve
         if (!requireNamespace("lme4")) {
           stop("This function needs `lme4` to be installed. Please install by running `install.packages('lme4')`.")
         }
-        adjusted <- residuals(lme4::lmer(paste(formula, formula_random), data = data))
+        model <- insight::get_residuals(lme4::lmer(paste(formula, formula_random), data = data))
       } else {
-        adjusted <- lm(formula, data = data)$residuals
+        model <- lm(formula, data = data)
       }
     }
+  }
+
+  adjusted <- insight::get_residuals(model)
+
+  # Re-add intercept if need be
+  if (keep_intercept) {
+    intercept <- insight::get_intercept(model)
+    if (length(intercept) > 1) intercept <- stats::median(intercept)  # For bayesian model
+    if (is.na(intercept)) intercept <- 0
+    adjusted <- adjusted + intercept
   }
 
   # Deal with missing data